Full RELAX Pipeline implementation
This is the first complete implementation of the data engine. README.md was updated to reflect the projects goals and instructions to run. Improved the `log_plot()` function to better inspect recordings. `testing.py` was added as a complete test file for the pipeline. MATLAB workspace was added to keep all MATLAB files and utilities in one place.